project fully loaded and ready for release

+++ lib/CategoryPages/JewelryCategoryPage.dart
@@ -170,7 +170,7 @@ class _SingleProductState extends State<SingleProduct> {
           });
           double price;
           if (widget.prd.discount == 'false')
-            price = widget.prd.price;
+            price = widget.prd.price.toDouble();
           else
             price = double.parse(widget.prd.newPrice);
           await _firestore
