+++ lib/components/Favorites.dart
@@ -146,7 +146,7 @@ class _SingleFavoritesProductState extends State<SingleFavoritesProduct> {
           });
           double price;
           if (widget.prd.discount == 'false')
-            price = widget.prd.price;
+            price = widget.prd.price.toDouble();
           else
             price = double.parse(widget.prd.newPrice);
           await _firestore
